Inline images as base64
