+++ Karta_postaci.md
@@ -1,0 +1,10 @@
+# Karta Postaci
+
+Ten plik zastępuje poprzednią binarną wersję `Karta postaci.docx`. Zawiera skrócone informacje opisujące główną bohaterkę gry.
+
+- **Imię:** Lilith
+- **Rasa:** Sukkub
+- **Etap początkowy:** 1
+- **Opis:** Demoniczna istota przyzwana przez gracza. W trakcie rozgrywki stopniowo pogłębia swoją korupcję i lojalność wobec przyzywającego.
+
+Plik ma służyć jedynie jako prosty punkt odniesienia podczas testów i będzie rozwijany w kolejnych wersjach projektu.
