+++ assets.js
@@ -1,0 +1,2 @@
+export const PLACEHOLDER_IMG = 'data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAAEAAAABCAYAAAAfFcSJAAAADElEQVQI12P4//8/AAX+Av5BjR+dAAAAAElFTkSuQmCC';
+export const FAVICON_DATA = 'data:image/x-icon;base64,AAABAAEAEBAQAAEABAAoAQAAFgAAACgAAAAQAAAAIAAAAAEABAAAAAAAgAAAAAAAAAAAAAAAEAAAAAAAADrL6YA7XlcAOwo+gAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAERERIiIiIRERESIgAAAiERERIAAAAAIRESIiAREQAgERIiIiERACARIgAAIRAAIBEiACIhEAIhESIgIiAAIhERESIiEAIhERESIiIAIgARERIiIQAiAAERERERAAIgAREREREQAAABEREREREAACERERERERESIREREREREREREAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA';
