+++ data/diaryEntries.js
@@ -1,0 +1,19 @@
+export const diaryEntries = [
+    { id: 'd_summoned_shock', title: "Pierwszy Dzień w Nowym Świecie", text: "Nie mogę uwierzyć, że to się naprawdę stało. Jeszcze wczoraj siedziałam w bibliotece Akademii, studiując teorię przyzwań, a dzisiaj... jestem tu. W innym wymiarze. Z nim. Przyzywaczem. Mój pierwszy prawdziwy przyzywacz! Profesor Azmodeus zawsze powtarzał, że pierwsze przyzwanie jest jak pierwszy pocałunek - nigdy się go nie zapomina. Na Asmodeusza, ja nawet nie wiem jak smakuje pocałunek, chyba, że liczyć te na zajęciach, z pozostałymi uczennicami akademii...", unlockConditions: { completedDialogueId: 'summoning_ritual' } },
+    { id: 'd_academic_pride_reflection', title: "Duma Akademicka... i Niepewność", text: "Powiedziałam mu o akademii... Poczułam się przez chwilę ważna, jak wtedy, gdy odbierałam dyplom. Ale potem... te jego pytania o praktykę. Zakuło. Może ma rację? Może teoria to nie wszystko?", unlockConditions: { completedDialogueId: 'academic_pride' } },
+    { id: 'd_first_touch_diary', title: "Dotyk, który Zapłonął", text: "Jego dłoń... tak blisko. A potem ten przypadkowy (czy na pewno?) dotyk. Przeszył mnie dreszcz, jakiego nie znałam. Podręczniki opisywały reakcje fizjologiczne, ale to... to było coś więcej. Bałam się i chciałam więcej jednocześnie. Co się ze mną dzieje?", unlockConditions: { completedDialogueId: 'first_touch_tension' } },
+    { id: 'd_vocal_thoughts_unleashed', title: "Myśli na Głos", text: "Powiedziałam mu. Powiedziałam, że od teraz będzie słyszał moje myśli. To było... wyzwalające. I przerażające. Czy on naprawdę chce wiedzieć, co kłębi się w głowie sukkuba? Co jeśli go odstraszę? Albo... co jeśli mu się to spodoba?", unlockConditions: { completedDialogueId: 'vocal_breakthrough_dialogue', requiresFlag: 'lilith_vocal_system_unlocked' } },
+    { id: 'd_first_kiss_diary', title: "Pocałunek Inny Niż Wszystkie", text: "Zrobiłam to. Pocałowałam go. Albo on mnie? To nie było jak te ćwiczenia na manekinach w akademii. To było... prawdziwe. Chaotyczne, niezdarne, ale... moje serce biło jak szalone. Chcę znowu.", unlockConditions: { completedDialogueId: 'first_real_kiss', requiresFlag: 'first_real_kiss_completed' } },
+    { id: 'd_true_desire_confessed', title: "Wyznanie", text: "W końcu to powiedziałam. Czego pragnę. Jego. Nie tylko jego esencji. Jego całego. Czy to miłość? Sukkuby nie kochają... prawda? Ale to, co czuję, przekracza wszystko, czego mnie uczono. Oddałam mu się całkowicie. I nigdy nie czułam się bardziej sobą.", unlockConditions: { completedDialogueId: 'breaking_point_true_desire', requiresFlag: 'first_vaginal_sex_completed' } },
+    { id: 'd_failed_technique', title: "Technika Numer 7 - Totalna Porażka", text: "Chcę umrzeć. Albo przynajmniej zapaść się pod ziemię...", unlockConditions: { completedDialogueId: 'academic_technique_fail', choiceMadeInDialogue: {dialogueId: 'academic_technique_fail', optionId: 'atf_laugh'} } },
+    { id: 'd_first_real_desire', title: "Pragnę Go", text: "Nie mogę spać. Leżę w łóżku i myślę tylko o nim...", unlockConditions: { stageRequired: 3, requiresFlag: 'dream_confession_completed' } },
+    { id: 'd_perfect_unity', title: "Jesteśmy Jednym", text: "Nie ma już Lilith-studentki... Jestem czymś więcej. Jesteśmy czymś więcej.", unlockConditions: { completedDialogueId: 'perfect_union_event_completed' } },
+    { id: 'd_touch_gentle', title: "Pierwszy Dotyk: Ciepło", text: "Jego dłoń... była taka ciepła. Delikatna. Czułam, jak moje serce przyspiesza. To było... miłe.", unlockConditions: { completedDialogueId: 'first_touch_tension', choiceMadeInDialogue: {dialogueId: 'first_touch_tension', optionId: 'ftt_subtle'} } },
+    { id: 'd_touch_rushed', title: "Pierwszy Dotyk: Zaskoczenie", text: "Tak nagle mnie przyciągnął... Jego dotyk był pewny, gorący. Zaskoczył mnie, ale... nie było to nieprzyjemne. Wręcz przeciwnie.", unlockConditions: { completedDialogueId: 'first_touch_tension', choiceMadeInDialogue: {dialogueId: 'first_touch_tension', optionId: 'ftt_bold'} } },
+    { id: 'd_theory_vs_reality_entry', title: "Wiedza a Rzeczywistość (Pamiętnik)", text: "Tyle lat wkuwania teorii... A rzeczywistość? Jest chaotyczna, nieprzewidywalna i... o wiele bardziej ekscytująca. Zwłaszcza z nim.", unlockConditions: { completedDialogueId: 'theory_vs_reality', stageRequired: 2} },
+    { id: 'd_curiosity_awakens_new', title: "Nowe Myśli: Przebudzenie Ciekawości", text: "Tyle pytań kłębi mi się w głowie... O niego. O ten świat. O siebie. Chcę wiedzieć więcej. Chcę doświadczać.", unlockConditions: { stageRequired: 2, requiresFlag: 'lilith_vocal_system_unlocked' } },
+    { id: 'd_kiss_sweet_new', title: "Pocałunek: Słodycz i Obietnica", text: "Jego usta... były takie miękkie. I smakowały obietnicą czegoś więcej. Czegoś, czego pragnę, nawet jeśli jeszcze tego nie rozumiem.", unlockConditions: { completedDialogueId: 'first_real_kiss', choiceMadeInDialogue: {dialogueId: 'first_real_kiss', optionId: 'frk_encourage'} } },
+    { id: 'd_temptation_rising_new', title: "Pokusa: Nocne Myśli i Płonące Ciało", text: "Nie mogę spać. Ciągle o nim myślę... O jego dotyku, o jego spojrzeniu. Moje ciało płonie. To uczucie... jest takie silne.", unlockConditions: { stageRequired: 4, requiresFlag: 'first_petting_completed' } },
+    { id: 'd_fear_of_essence_drain_new', title: "Lęk przed Esencją... i Utratą", text: "Im więcej o tym czytam i im więcej czuję... tym bardziej boję się, że mogę go skrzywdzić. Że moja natura weźmie górę. A ja... nie chcę go stracić.", unlockConditions: { stageRequired: 2, requiresFlag: 'lilith_is_fearful_of_harming' } }
+];
+
