+++ index.html
@@ -10,6 +10,7 @@
     <link href="https://fonts.googleapis.com/css2?family=Cinzel:wght@400;600;800&family=Crimson+Pro:ital,wght@0,400;0,600;1,400&display=swap" rel="stylesheet">
     <!-- Link to external CSS file -->
     <link rel="stylesheet" href="style.css">
+    <link rel="icon" type="image/x-icon" href="data:image/x-icon;base64,AAABAAEAEBAQAAEABAAoAQAAFgAAACgAAAAQAAAAIAAAAAEABAAAAAAAgAAAAAAAAAAAAAAAEAAAAAAAADrL6YA7XlcAOwo+gAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAERERIiIiIRERESIgAAAiERERIAAAAAIRESIiAREQAgERIiIiERACARIgAAIRAAIBEiACIhEAIhESIgIiAAIhERESIiEAIhERESIiIAIgARERIiIQAiAAERERERAAIgAREREREQAAABEREREREAACERERERERESIREREREREREREAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA">
 </head>
 <body class="bg-background text-text">
     <div class="max-w-screen-2xl mx-auto p-4 md:p-8 game-container">
@@ -53,7 +54,7 @@
                 <div id="lilith-card-panel" class="card">
                     <h2 class="card-header text-2xl md:text-3xl text-center" id="lilith-name-stage">Sukkub - Etap Przyzwana</h2>
                     <div class="w-full h-auto bg-surface-dark mx-auto mb-3 rounded-md flex items-center justify-center italic lilith-image-placeholder float-animation" id="lilith-image">
-                        <img id="lilith-img-tag" src="images/lilith_placeholder.png" alt="Obraz Lilith - kliknij, aby wygenerować esencję" class="w-full h-full object-cover rounded-md interactive-lilith-image">
+                        <img id="lilith-img-tag" src="data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAAEAAAABCAYAAAAfFcSJAAAADElEQVQI12P4//8/AAX+Av5BjR+dAAAAAElFTkSuQmCC" alt="Obraz Lilith - kliknij, aby wygenerować esencję" class="w-full h-full object-cover rounded-md interactive-lilith-image">
                     </div>
                     <p id="lilith-description" class="mt-2 italic text-center text-text-muted">Lilith jest w szoku po przyzwaniu...</p>
                     <p id="lilith-vocal-thought-display" class="text-center"></p>
@@ -85,7 +86,7 @@
                 <div id="diary-container" class="card">
                     <h3 class="card-header flex justify-between items-center">
                         <span>Pamiętnik Lilith</span>
-                        <img id="diary-toggle-icon" src="images/diary_icon.png" alt="Pokaż/Ukryj" class="diary-toggle-icon">
+                        <img id="diary-toggle-icon" src="data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAAEAAAABCAYAAAAfFcSJAAAADElEQVQI12P4//8/AAX+Av5BjR+dAAAAAElFTkSuQmCC" alt="Pokaż/Ukryj" class="diary-toggle-icon">
                     </h3>
                     <div id="diary-entries-list" class="space-y-2 hidden"></div>
                     <div id="diary-entry-content-area-custom" class="hidden">
