+++ style.css
@@ -273,3 +273,11 @@ h1, h2, h3, h4, h5, h6 {
 .scrollable-content::-webkit-scrollbar-thumb:hover {
     background: rgba(var(--color-primary-light), 1);
 }
+
+.dialogue-icon {
+    width: 40px;
+    height: 40px;
+    object-fit: cover;
+    display: block;
+    margin-bottom: 4px;
+}
